% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{attachShinyJs}
\alias{attachShinyJs}
\title{attach shiny.min.js to html tag}
\usage{
attachShinyJs(.tag)
}
\arguments{
\item{.tag}{a shiny tag}
}
\value{
a shiny tag
}
\description{
attach shiny.min.js to html tag
}
