% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmltools.R
\name{browseTag}
\alias{browseTag}
\title{As browseURL but it browse a shiny.tag or html class object}
\usage{
browseTag(tag = .Last.value)
}
\arguments{
\item{tag}{A shiny.tag or html class object}
}
\value{

}
\description{
As browseURL but it browse a shiny.tag or html class object
}
\examples{
none
}
