% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmltools.R
\name{browseTag2}
\alias{browseTag2}
\title{As browseURL but it browse a shiny.tag or html class object in RStudio Viewer}
\usage{
browseTag2(tag = .Last.value)
}
\arguments{
\item{tag}{A shiny.tag or html class object}
}
\value{

}
\description{
As browseURL but it browse a shiny.tag or html class object in RStudio Viewer
}
\examples{
none
}
