% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figma2.R
\name{create_rfig}
\alias{create_rfig}
\title{Convert figma CSS in clipboard to an R environment object}
\usage{
create_rfig()
}
\value{
an environment with `create_cssFile`, `div` and `style` objects.
}
\description{
Convert figma CSS in clipboard to an R environment object
}
\examples{
none
}
