% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{createWidget}
\alias{createWidget}
\title{Scaffold widget from web ui module files}
\usage{
createWidget(widgetname, modulename = NULL)
}
\arguments{
\item{widgetname}{a character}

\item{modulename}{a character (default=NULL) representing "R/modulename.R" module file. If omit, will look for "R/widgetname.R" as the module file and change it to "R/widgetname_support.R" to avoid widget file name conflict.}
}
\value{
none
}
\description{
Scaffold widget from web ui module files
}
\examples{
none
}
