% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig.R
\name{Fig}
\alias{Fig}
\title{Initiator for fig}
\usage{
Fig(attachmentSrc = NULL)
}
\arguments{
\item{attachmentSrc}{A path to attachment files folder.}
}
\value{

}
\description{
Initiator for fig
}
\examples{
none
}
