% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figma.R
\name{figma_Css2Html}
\alias{figma_Css2Html}
\title{Translate clipboard figma css to html}
\usage{
figma_Css2Html(update_autolayout_margin = F)
}
\arguments{
\item{update_autolayout_margin}{default =F}
}
\value{

}
\description{
Translate clipboard figma css to html
}
\examples{
none
}
