% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmltools.R
\name{generate_assets_attachment_href}
\alias{generate_assets_attachment_href}
\title{Generate assets attachment href when dependency attachment is vector of folder names}
\usage{
generate_assets_attachment_href(dependency)
}
\arguments{
\item{dependency}{a htmlDepency where src=c(file...)}
}
\value{
a list
}
\description{
Generate assets attachment href when dependency attachment is vector of folder names
}
