% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmltools.R
\name{htmlDependency2}
\alias{htmlDependency2}
\title{Create dependency other than script and link in the head}
\usage{
htmlDependency2(name, version, ...)
}
\arguments{
\item{name}{as in htmlDependency}

\item{version}{as in htmlDependency}

\item{...}{other inputs other than src, script and style}
}
\value{

}
\description{
Create dependency other than script and link in the head
}
\examples{
dep_ios <- htmlDependency2(
  name="ios-app",
  version="0.0.1",
  meta = list(
    `apple-mobile-web-app-capable`="yes",
    `apple-mobile-web-app-status-bar-style`="black"
  )
)
}
