% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{make_svg_responsive}
\alias{make_svg_responsive}
\title{Make svg responsive}
\usage{
make_svg_responsive(svgfile)
}
\arguments{
\item{svgfile}{a filepath}
}
\value{

}
\description{
Make svg responsive
}
\examples{
none
}
