% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmltools.R
\name{translate_HTML_fromClipboard}
\alias{translate_HTML_fromClipboard}
\title{Take copy from clipboard and translate html codes to R codes and write to clipboard for paste}
\usage{
translate_HTML_fromClipboard(prefix = T)
}
\arguments{
\item{prefix}{a logical, default=T, the paste tag will start with tags$}
}
\value{

}
\description{
Take copy from clipboard and translate html codes to R codes and write to clipboard for paste
}
\examples{
none.
}
