% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmltools.R
\name{save_html2}
\alias{save_html2}
\title{save html with mobile dependency}
\usage{
save_html2(tag = .Last.value, file)
}
\arguments{
\item{tag}{a shiny tag}

\item{file}{a filepath}
}
\value{
none.
}
\description{
save html with mobile dependency
}
