% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon_tags.R
\name{tags_icon}
\alias{tags_icon}
\title{Generate tag functions for icon svg's saved under a certain folder}
\usage{
tags_icon(iconFolder)
}
\arguments{
\item{iconFolder}{A path to the svg folder}
}
\value{
a list of tag functions
}
\description{
Generate tag functions for icon svg's saved under a certain folder
}
\examples{
icons <- tags_icon("assets/icon")
tagList(
  icons$`btn-contact-off`(),
  icons$`btn-contact-on`(),
  icons$`btn-home-off`(),
  icons$`btn-home-on`()) -> tagIcons
tagIcons |> econWeb::browseTag()
}
